src/BlogApplication/requirements.txt
django>=3.2.14
djangorestframework
django-crispy-forms
crispy-bootstrap4